+++ src/presentation/controllers/transaction/delete-transaction-controller.ts
@@ -1,6 +1,6 @@
 import { DeleteTransaction } from "src/domain/usecases/transaction/delete-transaction";
-import { Controller } from "../protocols/controller";
-import { HttpRequest, HttpResponse } from "../protocols/http";
+import { Controller } from "../../protocols/controller";
+import { HttpRequest, HttpResponse } from "../../protocols/http";
 
 export class DeleteTransactionController implements Controller {
   constructor(private readonly deleteTransaction: DeleteTransaction) {}
