+++ src/presentation/controllers/transaction/get-total-amount-controller.ts
@@ -1,6 +1,6 @@
 import { GetTotalAmount } from "src/domain/usecases/transaction/get-total-amount";
-import { Controller } from "../protocols/controller";
-import { HttpRequest, HttpResponse } from "../protocols/http";
+import { Controller } from "../../protocols/controller";
+import { HttpRequest, HttpResponse } from "../../protocols/http";
 
 export class GetTotalAmountController implements Controller {
   constructor(private readonly getTotalAmount: GetTotalAmount) {}
